Added events count in chart title.
The chart title will now display the number of events in the viewport.

Furthermore, this fixes a regression where new data points would not be shown in the current viewport as the extreme max value was set to the time where the popup was open. This enforced the viewport not to show this new point since it was out of bounds. Now we simply pass max = null so that it always displays new data.

Fixed a regression where the index set on the series are used in addPoint, thus the added points where inverted (tabCount being windowCount and windowCount being tabCount).